Rewrote the clustering code. Much faster now.
This takes advantage of the fact that our data tends to be heavy on
duplicates. I think we'll have to return to this eventually but for now
I'm happy with the performance.

Some ideas I have are adding incoming points to existing clusters and
doing some preliminary grouping. It's helpful that our data is either
very dense or spread out.