Actually returning modules and meta from summary.
The long term plan in here is to take the unique modules map and look
for key collisions and then introduce prefixes on the keys to
disambiguate things.

I think a generalized disambigation will be necessary anyway, akin to
having modules with differening kind/manufacture/version with the same
keys, to avoid misinterpreting those. For now, I'm leaving this alone
and just returning all the modules.

This is one of the reasons for having MetaID on DataRow, for the
easier case we can look up the actual meta and compare and come up
with disambiguation strategies.

When we do a summary, we'll have to do this disambiguation before we
summarize, so that we end up with the expected keys. So in the case
that 'temp' appears twice in a summary we'd turn their keys into:

<water-module-id>.temp
<some-other-module>.temp

Or even:

<index-of-water-module-meta>.temp
<index-of-other-module-meta>.temp