Create module/sensor model from status replies.
This also adds a few columns. source_id is so we can have a unique
index on that and hardware_id for those insertions, meanwhile we keep
the index on hardware_id and model_record_id.

There's also an updated_at that we use to batch updated records and
that's the main way we pick the set to return now.