Add column for records containing the protocol buffer message.
Right now we're to the point where parsing JSON is causing a real
performance hit when we're displaying records. I suspect that parsing
the PB is way faster and this is an easy way to get that in. Not sure
how useful having the JSON has been since we don't really use PG to
pull parts out.