Disable aggregating handler on main ingestion.
This won't really do everything we need, since we're only doing a file
at a time and many times they're uploaded at an interval on the hour
or less, so the way these are written they won't actually create
aggregates for those intervals.

They either need to be smarter, saving their state... or we need to
process the ingested sample data in bulk.

We can speed that up by running the aggregation on only recently
acquired data.