Viz side project, me learning d3 and vuejs stuff.
This is the first commit of a day of me playing with d3 and vuejs to
get my bearings. My d3 is very rusty and this is where I'm at right
now. I've tried to isolate the d3 stuff into single components. I do
think there may be some redundancy in the vuejs vs data hierarchy but
this is also where I'm at. Where things are I think it's isolated
enough to put a stake in the ground.

My approach has been to copy styling as much as I can from the
previous components and to make things more d3 friendly.