Revamp project photos to allow for etag based cache control.
This is also a proof of concept of this pattern. I wish I could have
avoided the use of the custom encoder but I simply couldn't. It, and
the awkwardness of the preventable write header make this solution
kind of weak. I do enjoy the separation it provides, though which is
the only reason I'm gonna go along.