Recreate project_and_station_activity view.
For some reason, my view was broken locally. Which makes me nervous. This
started when activity_test began to fail because no activity was being
returned. Hours of digging nudged me to recreate the view and things just
worked.

It seemed like the view as only returning the project activities.