Delete removed aggregates when processing completely.
This takes advantage of the nsamples row in the aggregates to prune
rows that should go away. Before processing a station completely, we
assign 0 to all the rows nsample columns. When we a upserting, this
column is updated and should never be <1, or the row should be gone.
After, we delete all rows with nsamples = 0 for that station.

This should handle all the main edge cases for getting rid of
aggregates that disappear.