feat: Separate sensors by modules in aggregated data.
A little easier than I expected, though this will require churning
over all the data again, so better now than alter.

We added a module_id column to the sensor data so that we can
differentiate between two modules on the same station. Still can't
believe I designed and wrote this redesign without thinking of that.

From there everything works as expected, we consider a 'sensor' to be
a two item array of numbers: [module-id, sensor-id] so any place where
we had an array of sensor IDs has become an array of these
arrays. This slides nicely into the bookmark, which I've broken
backwards compatibility, as things are very early on.

The only remaining thing is to reimport all of the data.