fix: Revamp time labels to remove repetition.
This changes the time axis labels to basically only include the date
part if the date part isn't unique and to do the same for time, but
never for both at the same... time.

So you only see the date if the times are all the same and vice versa.