feat: Query by token instead of module identifiers.
A module token is a shorthand for referring to multiple modules of the
same kind. In this case, we're using this to specifically refer to the
same module over the lifetime of the station.

Put another way. Module identifiers of the same module token can be
queried together to return a coherent graph. It probably needs a
better name.