feat: Node backend for server-side chart rendering of time series.
This effectively acts a headless browser. Somebody requests a rendered chart by
passing the bookmark for that chart and the backend will issue the same HTTP
queries the browser would to retrieve the data. That is then passed onto Vega
for rendering.

This means overtime we can decide how bookmarks with multiple charts are
displayed and we can also "narrow" bookmarks to only include the chart we'd like
for more control.