refact: Assumptions in notification querying.
This first removes some unused query methods, which I love to remove because
we're maintaining less SQL than we were before.

The QueryByUserIDWithPost approach works right now because all notifications are
discussion related. Instead of sticking with that I'd rather make that
assumption inside the query method itself and extend that when we add more
notification types.

For example, when we add a new type of notification we can decide to "widen" the
query to the notifications table, adding LEFT JOINs or do separate/parallel
queries for the notifications and their details.

This does leave the awkward map/dict handling we have when the notification is
first created. I had originally hoped to include a JSON payload in the `body`
column, which contained enough information to make queries against the
discussion tables unnecessary, may be too late to make that change, though and
that kind of optimization may be a bad idea for small teams, esp while
performance is fine.