fix(perf): Rely on nsamples = 0 deletion to remove stale samples.
This was redundantly deleting rows that were "removed" during the aggregation
pass, which is pretty rare to begin with. So, instead of doing the very frequent
DELETE by time we just rely on the DELETE that removes readings with 0 samples,
which we setup before we begin the aggregation.

There's still more here to fix performance wise but I also hope to trial an
influxdb backend, soon.