fix: Add "tail" threshold w/o a label for catch-all to show all data.
Prior to this change, if any data appeared in the data set that was out of
range, it wouldn't be visible. This is kind of annoying because then we're
forced to create a label for this data and it may not be what we'd like. So,
instead we allow a last ditch level to be added w/o a label so the data shows
up. There may be other better ways to do this but this was a surprise.