feat: Cap time-ranged queries with enveloping values.
This is one of those things that's convincing me to try throwing together a
quick InfluxDB backend.

The issue is when querying for a range of time where no values occur, so the
returned rows are empty. Only, what if before the range the last value was 100
and after it's something else, shouldn't we see the line?