fix: Show associated stations, faster viz load time.
Setting the "visible" range to either infinity and then the actual
range of the data was causing the page to reload twice. So now we treat
the inifnite range as the dated range to avoid the double refresh.

Also, we weren't including stations in an updated bookmark in the
active workspace.