fix: Remove busy when showing cached data.
I think that this behavior needs work, but for now I prefer this to how
things were before. Always showing the busy is annoying in situations
where a click is doing nothing, like double clicking on an already
zoomed out view. Instead we should be diff'ing the bookmark and using
that to determine if we should.