feat: Attach distance from first station on associated query.
This won't scale and will need to be refactored when I've got more time. I spent at least two hours trying to get a fancy solution working that involved a temporary table so that we could start to DRY up the StationFull querying and ran into all kinds of trouble with Pg refusing to find my temporary tables.