refact: Station independent associated query.
Starting to work on cleaning up how we handle metadata in viz and this is the first thing to change. Instead of returning a very station-specific view of the associated stations, we return a "station independent" view of the relationships. This is way less confusing than before but also will require significant changes in the browser. It is also one step closer to use being able to remove the extra calls to get sensor information.