fix(viz): Sparse data querying and adjusting bar chart bin sizes.
This is a move towards addressing the weird overflow behavior, which happens
when the bins end up wider than the queried time range. Say a 1 hour or even 1
minute bar chart bin when the time range is less than a minute. So to fix that
we adjust it to use seconds when the range is that short.

Another change is to check the total number of samples for one that is
disportionately small for the queried range so we can return finer grained data.

These may require tweaking, still trying to figure out what will work long term.