fix: Use fixed ending date for fast zoom links.
This effectively reverts commit 05eafdf421745ef8a37583c4a894353eb0006867, which
had started using Time.Max. From the comment I seem to have thought this might
be better done elsewhere, and that may still be true. For example there may be
scrubber issues as a result of this. For now I'm going back to this.