fix(viz): Truncate max end time to current system time.
The Eternity check is here in the TsDB querying so that we know to avoid doing
date math on the given times in order to estimate the sample size. This also
won't work if the end time of the query is the 'maximum end time'. We fix this
by truncating that to the current system time.