fix(viz): Buffer queried data to avoid double or even triple refreshes.
We wait until all the data has come back in before doing the final resolve and
sending the data back to the viz layer for processing. Without these we were
seeing stuttering as charts would refresh after a second data query was
returned.

Ideally this would still work at the chart level, so each chart could show as
the data for that chart came in.