refact: Use module_sensor for sensor times instead of old table.
No longer maintaining the aggregated_sensor_updated table after
the webhook work we've been doing. This now uses the single time
on module_sensor and does an awkward join on the full_sensor_key
to get the aggregated_sensor.id, which is still better in IMO
than how things were before and less confusing to dig around in.

This requires the previous visible_configuration commit.