feat(wip): Improved drag handling behavior.
Some notable differences include using `clip` with a custom SVG path to hide values outside of the time domain, which still needs some size tuning.

Also, all updates are internal, and signal hooks are used to move the scrubber. This also fixes several issues with dragging outside of the view.

Learned a lot from this:
https://vega.github.io/vega/examples/zoomable-scatter-plot/