feat(viz): Prevent dragging past end of the data.
This required the server to return the real end of the data. This refactors the queryRanges method to make that easier. One notable change is to always query for the data min/max ranges, ignoring the queried start/end. We can do this because when the queried start/end are real dates, we use them and can ignore
the ranges. The ranges are only used right now when one of the parameters is infinite and should catch "all the data" anyway.