refact: Migrate to a new, maintained que-go library.
Notice the table name and schema has changed!

This also appears to fix a deadlock I was seeing with the older
library when increasing the number of works.

EXCEPT, there were some deadlock induced retries on the SensorDataBatch
messages and things took a little longer. At least things don't hang
forever, though. Gonna test as this could also be due to the upgrade to
v4 of jackc/pgx/v5