feat: Wrap ingestion in a saga.
This needs a ton of work, it's getting the job done though. Some things that
need attention are how clean the handlers are and the ceremony around
updating/saving the saga state.

I also think some work needs to be done to make updating the sagas more
infrequent. Right now the optimistic locking fails a lot on the sensor data
batch committed messages.