fix(txs): Unwrap before scope, always enlist in dbpool.
This also drops the field level ctx from MessageContext and requires
the caller to provide one. Now that we unwrap TransportMessage's before
transactions the context provided to MessageContext was missing the
scope.

With this change we can rollback the handler and then decide where
the message should go after and also ensure that any operations
that are tx-aware are enlisted.