Build our docker-compose stacks from images.
We deploy 'stacks' of docker containers for each of our services. These
are tar archives that can be rsync'd to servers and then very simple
scripts manage the task of uncompressing them and rolling forward.

I would love to eventually try a simplified fleet management tool for
handling this has served us very well for a while.