Use returned data end / last sensor reading.
Because of aggregating the time of the last sensor reading sample
that's returned may not be the actual time of the reading and
instead rounded to its nearest bucket.

This changes the fast zoom behavior to always use the `dataEnd`
value that's returned with every query, rather than the `_max(time)`
of the data set.